Packaged the tests app
Package the tests app to be used as a reusable Django app;
Added settings and updated urls for tests app so that it can be tested without a wrapping project;
Added setuptools files such as CHANGES, LICENSE, README, MANIFEST and setup.py.
Added tox.ini for Tox integration;
Updated .gitignore to ignore package files and folders which should not be committed;
Removed all previous folders and files, specifically the training project, the tutorial polls app and some global templates.